+++ Changelog.md
@@ -3,6 +3,7 @@
 Features:
 
 Bugfixes:
+ * Code Generator: Provide "new account gas" for low-level ``callcode`` and ``delegatecall``.
 
 ### 0.4.14 (2017-07-31)